third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/unloading-documents/003-expected.txt

This is a testharness.js-based test.
[FAIL] document.open in beforeunload with link
  assert_equals: expected "0123456789" but got "01234568"
Harness: the test ran to completion.
